make utility function public
